+++ src/contexts/UserContext.tsx
@@ -1,10 +1,10 @@
 /* eslint-disable @typescript-eslint/no-empty-function */
-import React, { useState, useEffect, useContext, useRef} from "react";
+import React, { useState, useEffect, useContext} from "react";
 import { auth } from "../firebase/firebase";
 import { User, onAuthStateChanged } from "firebase/auth";
 import UpdateCssVariablePaletteObject from "../components/HelperComponents/UpdateCssVariablePaletteObject";
 import { UserDocument } from "../firebase/firestoreDocumentInterfaces";
-import { getUserFromUserId } from "../firebase/firestoreGet";
+import { getUserFromUserId } from "../firebase/GET/userGets";
 
 interface UserInfo {
     user: User | null,
